Refactored tests to meet reqirements
