+++ csharp-advanced-topics/ActionAndFuncDelegatesInCSharp/ActionAndFuncDelegatesInCSharp/FuncDelegates.cs
@@ -5,10 +5,10 @@ public class FuncDelegates
     {
 
         Func<string, string, string> greetWithNameFirst = new Func<string, string, string>(GreetWithName);
-        Console.WriteLine(greetWithNameFirst("Ralph", "Nyoni"));
+        //Console.WriteLine(greetWithNameFirst("Ralph", "Nyoni"));
 
         Func<string, string, string> greetWithNameSecond = GreetWithName;
-        Console.WriteLine(greetWithNameSecond("Nyoni", "Ralph"));
+        //Console.WriteLine(greetWithNameSecond("Nyoni", "Ralph"));
     }
     public static string GreetWithName(string firstName, string lastName)
     {
